requirements.txt
altgraph==0.17.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
bottle==0.13.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clr_loader==0.2.7.post0
colorama==0.4.6
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
idna==3.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
proxy_tools==0.1.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
python-dateutil==2.9.0.post0
pythonnet==3.0.5
pywebview==5.4
pywin32==310
pywin32-ctypes==0.2.3
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.14.0
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
